Moved ProcessIndirectFitParameter related Files to Workflow
refs #13514
